+++ vkconfig_core/environment.cpp
@@ -77,8 +77,8 @@ static const char* GetActiveDefault(Active active) {
     assert(active >= ACTIVE_FIRST && active <= ACTIVE_LAST);
 
     static const char* table[] = {
-        "Validation - Standard",  // ACTIVE_CONFIGURATION
-        ""                        // ACTIVE_EXECUTABLE
+        "Validation",  // ACTIVE_CONFIGURATION
+        ""             // ACTIVE_EXECUTABLE
     };
     static_assert(countof(table) == ACTIVE_COUNT, "The tranlation table size doesn't match the enum number of elements");
 
@@ -130,7 +130,7 @@ std::string GetLoaderDebugToken(LoaderMessageLevel level) {
 
 Environment::Environment(PathManager& paths, const Version& api_version)
     : api_version(api_version),
-      loader_message_level(LAODER_MESSAGE_NONE),
+      loader_message_level(GetLoaderDebug(qgetenv("VK_LOADER_DEBUG").toStdString())),
 // Hack for GitHub C.I.
 #if VKC_PLATFORM == VKC_PLATFORM_WINDOWS && (QT_VERSION >= QT_VERSION_CHECK(5, 10, 0))
       running_as_administrator(IsUserAnAdmin()),
@@ -221,6 +221,34 @@ void Environment::Reset(ResetMode mode) {
             Set(ACTIVE_CONFIGURATION, "Validation");
             break;
         }
+        case CLEAR: {
+            this->Reset(DEFAULT);
+
+            QSettings settings;
+            settings.setValue("VKCONFIG_WIDGET_SETTING_INT", false);
+            settings.setValue("VKCONFIG_WIDGET_SETTING_FLOAT", false);
+            settings.setValue("VKCONFIG_WIDGET_SETTING_FRAMES", false);
+            settings.setValue("VKCONFIG_WARN_MISSING_LAYERS_IGNORE", false);
+
+            settings.setValue("overrideActive", false);
+            settings.setValue("applyOnlyToList", false);
+            settings.setValue("keepActiveOnExit", false);
+
+            settings.setValue("restartWarning", false);
+            settings.setValue("warnAboutShutdownState", false);
+
+            settings.setValue(VKCONFIG_KEY_EXIT, false);
+            settings.setValue(VKCONFIG_KEY_RESTART, false);
+
+            const std::string loader_debug_message(qgetenv("VK_LOADER_DEBUG"));
+            if (loader_debug_message.empty()) {
+                loader_message_level = LAODER_MESSAGE_NONE;
+            } else {
+                loader_message_level = GetLoaderDebug(loader_debug_message);
+            }
+            settings.setValue(VKCONFIG_KEY_LOADER_MESSAGE, static_cast<int>(loader_message_level));
+            break;
+        }
         case SYSTEM: {
             const bool result_env = Load();
             assert(result_env);
@@ -241,12 +269,11 @@ bool Environment::Load() {
     QSettings settings;
 
     // Load "first_run"
-    first_run = settings.value(VKCONFIG_KEY_INITIALIZE_FILES, QVariant(first_run)).toBool();
+    first_run = settings.value(VKCONFIG_KEY_INITIALIZE_FILES, first_run).toBool();
 
     // Load "version": If the version doesn't exist of it's an old version of vkconfig
     const Version default_version(first_run || !SUPPORT_LAYER_CONFIG_2_0_0 ? version : Version("2.0.0"));
-    version =
-        Version(settings.value(VKCONFIG_KEY_VKCONFIG_VERSION, QVariant(default_version.str().c_str())).toString().toStdString());
+    version = Version(settings.value(VKCONFIG_KEY_VKCONFIG_VERSION, default_version.str().c_str()).toString().toStdString());
 
     if (version == Version("2.0.0")) {  // The version is an old development version, unknown and unsupported, hard reset.
         Reset(DEFAULT);
@@ -255,9 +282,9 @@ bool Environment::Load() {
 
     // Load 'override_mode"
     if (SUPPORT_LAYER_CONFIG_2_0_1 && version <= Version("2.0.1")) {
-        const bool active = settings.value("overrideActive", QVariant(first_run)).toBool();
-        const bool active_list = settings.value("applyOnlyToList", QVariant(first_run)).toBool();
-        const bool active_persistent = settings.value("keepActiveOnExit", QVariant(first_run)).toBool();
+        const bool active = settings.value("overrideActive", first_run).toBool();
+        const bool active_list = settings.value("applyOnlyToList", first_run).toBool();
+        const bool active_persistent = settings.value("keepActiveOnExit", first_run).toBool();
         if (!active)
             override_state = OVERRIDE_STATE_DISABLED;
         else if (active_list && active_persistent)
@@ -277,13 +304,8 @@ bool Environment::Load() {
     }
 
     // Load loader debug message state
-    const std::string loader_debug_message(qgetenv("VK_LOADER_DEBUG"));
-    if (loader_debug_message.empty()) {
-        loader_message_level =
-            static_cast<LoaderMessageLevel>(settings.value(VKCONFIG_KEY_LOADER_MESSAGE, QVariant(loader_message_level)).toInt());
-    } else {
-        loader_message_level = GetLoaderDebug(loader_debug_message);
-    }
+    loader_message_level = static_cast<LoaderMessageLevel>(
+        settings.value(VKCONFIG_KEY_LOADER_MESSAGE, static_cast<int>(loader_message_level)).toInt());
 
     // Load active configuration
     for (std::size_t i = 0; i < ACTIVE_COUNT; ++i) {
@@ -381,7 +403,7 @@ bool Environment::Save() const {
     settings.setValue(VKCONFIG_KEY_OVERRIDE_MODE, override_state);
 
     // Save 'loader_message'
-    settings.setValue(VKCONFIG_KEY_LOADER_MESSAGE, loader_message_level);
+    settings.setValue(VKCONFIG_KEY_LOADER_MESSAGE, static_cast<int>(loader_message_level));
 
     // Save active state
     for (std::size_t i = 0; i < ACTIVE_COUNT; ++i) {