+++ vkconfig/CHANGELOG.md
@@ -32,6 +32,7 @@
 - Fix Vulkan application crash when updating the Vulkan loader
 - Fix User-Defined layer path window 'Ok' button being disabled incorrectly
 - Fix blank setting tree after using the user-defined layer path window #1377
+- Fix 'Reset To Default' than wasn't bringing back notification pop-ups.
 
 ## [Vulkan Configurator 2.2.0 for Vulkan SDK 1.2.170.0](https://github.com/LunarG/VulkanTools/tree/sdk-1.2.170.0) - March 2021